Replace BernoulliMLPRegressor
This commit replaces the old BernoulliMLPRegressor with the one
implemented using garage.tf.models
