+++ src/garage/tf/regressors/__init__.py
@@ -3,8 +3,6 @@ from garage.tf.regressors.base import StochasticRegressor
 from garage.tf.regressors.base2 import Regressor2
 from garage.tf.regressors.base2 import StochasticRegressor2
 from garage.tf.regressors.bernoulli_mlp_regressor import BernoulliMLPRegressor
-from garage.tf.regressors.bernoulli_mlp_regressor_with_model import (
-    BernoulliMLPRegressorWithModel)
 from garage.tf.regressors.categorical_mlp_regressor import (
     CategoricalMLPRegressor)
 from garage.tf.regressors.continuous_mlp_regressor import (
@@ -15,7 +13,7 @@ from garage.tf.regressors.gaussian_cnn_regressor_model import (
 from garage.tf.regressors.gaussian_mlp_regressor import GaussianMLPRegressor
 
 __all__ = [
-    'BernoulliMLPRegressor', 'BernoulliMLPRegressorWithModel',
+    'BernoulliMLPRegressor',
     'CategoricalMLPRegressor', 'ContinuousMLPRegressor',
     'GaussianCNNRegressor', 'GaussianCNNRegressorModel',
     'GaussianMLPRegressor', 'Regressor2', 'StochasticRegressor',
