+++ src/garage/tf/regressors/bernoulli_mlp_regressor.py
@@ -1,167 +1,242 @@
+"""Bernoulli MLP Regressor based on MLP with Normalized Inputs."""
 from dowel import tabular
 import numpy as np
 import tensorflow as tf
 
-from garage.core import Serializable
-from garage.tf.core import LayersPowered, Parameterized
-import garage.tf.core.layers as L
-from garage.tf.core.network import MLP
 from garage.tf.distributions import Bernoulli
 from garage.tf.misc import tensor_utils
+from garage.tf.models import NormalizedInputMLPModel
 from garage.tf.optimizers import ConjugateGradientOptimizer, LbfgsOptimizer
+from garage.tf.regressors import StochasticRegressor2
 
 
-class BernoulliMLPRegressor(LayersPowered, Serializable, Parameterized):
+class BernoulliMLPRegressor(StochasticRegressor2):
     """
-    A class for performing regression (or classification, really) by fitting a
-    bernoulli distribution to each of the output units.
+    BernoulliMLPRegressor with garage.tf.models.NormalizedInputMLPModel.
+
+    A class for performing regression by fitting a Bernoulli distribution to
+    the outputs.
+
+    Args:
+        input_shape (tuple[int]): Input shape of the training data. Since an
+            MLP model is used, implementation assumes flattened inputs. The
+            input shape of each data point should thus be of shape (x, ).
+        output_dim (int): Output dimension of the model.
+        name (str): Model name, also the variable scope.
+        hidden_sizes (list[int]): Output dimension of dense layer(s) for
+            the MLP for the network. For example, (32, 32) means the MLP
+            consists of two hidden layers, each with 32 hidden units.
+        hidden_nonlinearity (callable): Activation function for intermediate
+            dense layer(s). It should return a tf.Tensor. Set it to
+            None to maintain a linear activation.
+        hidden_w_init (callable): Initializer function for the weight
+            of intermediate dense layer(s). The function should return a
+            tf.Tensor. Default is Glorot uniform initializer.
+        hidden_b_init (callable): Initializer function for the bias
+            of intermediate dense layer(s). The function should return a
+            tf.Tensor. Default is zero initializer.
+        output_nonlinearity (callable): Activation function for output dense
+            layer. It should return a tf.Tensor. Set it to None to
+            maintain a linear activation.
+        output_w_init (callable): Initializer function for the weight
+            of output dense layer(s). The function should return a
+            tf.Tensor. Default is Glorot uniform initializer.
+        output_b_init (callable): Initializer function for the bias
+            of output dense layer(s). The function should return a
+            tf.Tensor. Default is zero initializer.
+        optimizer (garage.tf.Optimizer): Optimizer for minimizing the negative
+            log-likelihood. Defaults to LbsgsOptimizer
+        optimizer_args (dict): Arguments for the optimizer. Default is None,
+            which means no arguments.
+        tr_optimizer (garage.tf.Optimizer): Optimizer for trust region
+            approximation. Defaults to ConjugateGradientOptimizer.
+        tr_optimizer_args (dict): Arguments for the trust region optimizer.
+            Default is None, which means no arguments.
+        use_trust_region (bool): Whether to use trust region constraint.
+        max_kl_step (float): KL divergence constraint for each iteration.
+        normalize_inputs (bool): Bool for normalizing inputs or not.
+        layer_normalization (bool): Bool for using layer normalization or not.
     """
 
-    def __init__(
-            self,
-            input_shape,
-            output_dim,
-            name='BernoulliMLPRegressor',
-            hidden_sizes=(32, 32),
-            hidden_nonlinearity=tf.nn.relu,
-            optimizer=None,
-            tr_optimizer=None,
-            use_trust_region=True,
-            max_kl_step=0.01,
-            normalize_inputs=True,
-            no_initial_trust_region=True,
-    ):
-        """
-        :param input_shape: Shape of the input data.
-        :param output_dim: Dimension of output.
-        :param hidden_sizes: Number of hidden units of each layer of the mean
-        network.
-        :param hidden_nonlinearity: Non-linearity used for each layer of the
-        mean network.
-        :param optimizer: Optimizer for minimizing the negative log-likelihood.
-        :param use_trust_region: Whether to use trust region constraint.
-        :param max_kl_step: KL divergence constraint for each iteration
-        """
-        Parameterized.__init__(self)
-        Serializable.quick_init(self, locals())
+    def __init__(self,
+                 input_shape,
+                 output_dim,
+                 name='BernoulliMLPRegressor',
+                 hidden_sizes=(32, 32),
+                 hidden_nonlinearity=tf.nn.relu,
+                 hidden_w_init=tf.glorot_uniform_initializer(),
+                 hidden_b_init=tf.zeros_initializer(),
+                 output_nonlinearity=tf.nn.sigmoid,
+                 output_w_init=tf.glorot_uniform_initializer(),
+                 output_b_init=tf.zeros_initializer(),
+                 optimizer=None,
+                 optimizer_args=None,
+                 tr_optimizer=None,
+                 tr_optimizer_args=None,
+                 use_trust_region=True,
+                 max_kl_step=0.01,
+                 normalize_inputs=True,
+                 layer_normalization=False):
 
-        with tf.compat.v1.variable_scope(name):
+        super().__init__(input_shape, output_dim, name)
+        self._use_trust_region = use_trust_region
+        self._max_kl_step = max_kl_step
+        self._normalize_inputs = normalize_inputs
+
+        with tf.compat.v1.variable_scope(self._name, reuse=False) as vs:
+            self._variable_scope = vs
+            if optimizer_args is None:
+                optimizer_args = dict()
+            if tr_optimizer_args is None:
+                tr_optimizer_args = dict()
 
             if optimizer is None:
-                optimizer = LbfgsOptimizer()
+                optimizer = LbfgsOptimizer(**optimizer_args)
+            else:
+                optimizer = optimizer(**optimizer_args)
+
             if tr_optimizer is None:
-                tr_optimizer = ConjugateGradientOptimizer()
+                tr_optimizer = ConjugateGradientOptimizer(**tr_optimizer_args)
+            else:
+                tr_optimizer = tr_optimizer(**tr_optimizer_args)
 
-            self.output_dim = output_dim
-            self.optimizer = optimizer
-            self.tr_optimizer = tr_optimizer
+            self._optimizer = optimizer
+            self._tr_optimizer = tr_optimizer
 
-            p_network = MLP(
-                input_shape=input_shape,
-                output_dim=output_dim,
-                hidden_sizes=hidden_sizes,
-                hidden_nonlinearity=hidden_nonlinearity,
-                output_nonlinearity=tf.nn.sigmoid,
-                name='p_network')
+        self.model = NormalizedInputMLPModel(
+            input_shape,
+            output_dim,
+            hidden_sizes=hidden_sizes,
+            hidden_nonlinearity=hidden_nonlinearity,
+            hidden_w_init=hidden_w_init,
+            hidden_b_init=hidden_b_init,
+            output_nonlinearity=output_nonlinearity,
+            output_w_init=output_w_init,
+            output_b_init=output_b_init,
+            layer_normalization=layer_normalization)
 
-            l_p = p_network.output_layer
+        self._dist = Bernoulli(output_dim)
 
-            LayersPowered.__init__(self, [l_p])
+        self._initialize()
 
-            xs_var = p_network.input_layer.input_var
-            ys_var = tf.compat.v1.placeholder(
-                dtype=tf.float32, shape=(None, output_dim), name='ys')
-            old_p_var = tf.compat.v1.placeholder(
-                dtype=tf.float32, shape=(None, output_dim), name='old_p')
+    def _initialize(self):
+        input_var = tf.compat.v1.placeholder(tf.float32,
+                                             shape=(None, ) +
+                                             self._input_shape)
 
-            x_mean_var = tf.compat.v1.get_variable(
-                name='x_mean',
-                initializer=tf.zeros_initializer(),
-                shape=(1, ) + input_shape)
-            x_std_var = tf.compat.v1.get_variable(
-                name='x_std',
-                initializer=tf.ones_initializer(),
-                shape=(1, ) + input_shape)
+        with tf.compat.v1.variable_scope(self._variable_scope):
+            self.model.build(input_var)
 
-            normalized_xs_var = (xs_var - x_mean_var) / x_std_var
+            ys_var = tf.compat.v1.placeholder(dtype=tf.float32,
+                                              name='ys',
+                                              shape=(None, self._output_dim))
 
-            p_var = L.get_output(l_p,
-                                 {p_network.input_layer: normalized_xs_var})
+            old_prob_var = tf.compat.v1.placeholder(dtype=tf.float32,
+                                                    name='old_prob',
+                                                    shape=(None,
+                                                           self._output_dim))
 
-            old_info_vars = dict(p=old_p_var)
-            info_vars = dict(p=p_var)
+            y_hat = self.model.networks['default'].y_hat
 
-            dist = self._dist = Bernoulli(output_dim)
+            old_info_vars = dict(p=old_prob_var)
+            info_vars = dict(p=y_hat)
 
-            mean_kl = tf.reduce_mean(dist.kl_sym(old_info_vars, info_vars))
+            mean_kl = tf.reduce_mean(
+                self._dist.kl_sym(old_info_vars, info_vars))
 
-            loss = -tf.reduce_mean(dist.log_likelihood_sym(ys_var, info_vars))
+            loss = -tf.reduce_mean(
+                self._dist.log_likelihood_sym(ys_var, info_vars))
 
-            predicted = p_var >= 0.5
+            predicted = y_hat >= 0.5
 
-            self.f_predict = tensor_utils.compile_function([xs_var], predicted)
-            self.f_p = tensor_utils.compile_function([xs_var], p_var)
-            self.l_p = l_p
+            self._f_predict = tensor_utils.compile_function([input_var],
+                                                            predicted)
+            self._f_prob = tensor_utils.compile_function([input_var], y_hat)
 
-            self.optimizer.update_opt(
+            self._optimizer.update_opt(loss=loss,
+                                       target=self,
+                                       network_output=[y_hat],
+                                       inputs=[input_var, ys_var])
+            self._tr_optimizer.update_opt(
                 loss=loss,
                 target=self,
-                network_outputs=[p_var],
-                inputs=[xs_var, ys_var])
-            self.tr_optimizer.update_opt(
-                loss=loss,
-                target=self,
-                network_outputs=[p_var],
-                inputs=[xs_var, ys_var, old_p_var],
-                leq_constraint=(mean_kl, max_kl_step))
-
-            self.use_trust_region = use_trust_region
-            self.name = name
-
-            self.normalize_inputs = normalize_inputs
-            self.x_mean_var = x_mean_var
-            self.x_std_var = x_std_var
-            self.first_optimized = not no_initial_trust_region
+                network_output=[y_hat],
+                inputs=[input_var, ys_var, old_prob_var],
+                leq_constraint=(mean_kl, self._max_kl_step))
+            self._first_optimized = False
 
     def fit(self, xs, ys):
-        if self.normalize_inputs:
+        """
+        Fit with input data xs and label ys.
+
+        Args:
+            xs (numpy.ndarray): Input data.
+            ys (numpy.ndarray): Label of input data.
+        """
+        if self._normalize_inputs:
             # recompute normalizing constants for inputs
-            new_mean = np.mean(xs, axis=0, keepdims=True)
-            new_std = np.std(xs, axis=0, keepdims=True) + 1e-8
-            tf.compat.v1.get_default_session().run(
-                tf.group(
-                    tf.compat.v1.assign(self.x_mean_var, new_mean),
-                    tf.compat.v1.assign(self.x_std_var, new_std),
-                ))
-            # self._x_mean_var.set_value(np.mean(xs, axis=0, keepdims=True))
-            # self._x_std_var.set_value(
-            #     np.std(xs, axis=0, keepdims=True) + 1e-8)
-        if self.use_trust_region and self.first_optimized:
-            old_p = self.f_p(xs)
-            inputs = [xs, ys, old_p]
-            optimizer = self.tr_optimizer
+            self.model.networks['default'].x_mean.load(
+                np.mean(xs, axis=0, keepdims=True))
+            self.model.networks['default'].x_std.load(
+                np.std(xs, axis=0, keepdims=True) + 1e-8)
+
+        if self._use_trust_region and self._first_optimized:
+            # To use trust region constraint and optimizer
+            old_prob = self._f_prob(xs)
+            inputs = [xs, ys, old_prob]
+            optimizer = self._tr_optimizer
         else:
             inputs = [xs, ys]
-            optimizer = self.optimizer
+            optimizer = self._optimizer
         loss_before = optimizer.loss(inputs)
-        if self.name:
-            prefix = self.name + '/'
-        else:
-            prefix = ''
-        tabular.record(prefix + 'LossBefore', loss_before)
+        tabular.record('{}/LossBefore'.format(self._name), loss_before)
         optimizer.optimize(inputs)
         loss_after = optimizer.loss(inputs)
-        tabular.record(prefix + 'LossAfter', loss_after)
-        tabular.record(prefix + 'dLoss', loss_before - loss_after)
-        self.first_optimized = True
+        tabular.record('{}/LossAfter'.format(self._name), loss_after)
+        tabular.record('{}/dLoss'.format(self._name), loss_before - loss_after)
+        self._first_optimized = True
 
     def predict(self, xs):
-        return self.f_predict(np.asarray(xs))
+        """
+        Predict ys based on input xs.
 
-    def sample_predict(self, xs):
-        p = self.f_p(xs)
-        return self._dist.sample(dict(p=p))
+        Args:
+            xs (numpy.ndarray): Input data.
 
-    def predict_log_likelihood(self, xs, ys):
-        p = self.f_p(np.asarray(xs))
-        return self._dist.log_likelihood(np.asarray(ys), dict(p=p))
+        Return:
+            The predicted ys (one hot vectors).
+        """
+        return self._f_predict(xs)
+
+    def log_likelihood_sym(self, x_var, y_var, name=None):
+        """
+        Symbolic graph of the log likelihood.
+
+        Args:
+            x_var (tf.Tensor): Input tf.Tensor for the input data.
+            y_var (tf.Tensor): Input tf.Tensor for the one hot label of data.
+            name (str): Name of the new graph.
+
+        Return:
+            tf.Tensor output of the symbolic log likelihood.
+        """
+        with tf.compat.v1.variable_scope(self._variable_scope):
+            prob, _, _ = self.model.build(x_var, name=name)
+
+        return self._dist.log_likelihood_sym(y_var, dict(p=prob))
+
+    def get_params_internal(self, **args):
+        """Get the params, which are the trainable variables."""
+        return self._variable_scope.trainable_variables()
+
+    def __getstate__(self):
+        """Object.__getstate__."""
+        new_dict = super().__getstate__()
+        del new_dict['_f_predict']
+        del new_dict['_f_prob']
+        return new_dict
+
+    def __setstate__(self, state):
+        """Object.__setstate__."""
+        super().__setstate__(state)
+        self._initialize()
