+++ tests/garage/tf/regressors/test_bernoulli_mlp_regressor.py
@@ -6,7 +6,7 @@ import pytest
 import tensorflow as tf
 
 from garage.tf.optimizers import ConjugateGradientOptimizer, LbfgsOptimizer
-from garage.tf.regressors import BernoulliMLPRegressorWithModel
+from garage.tf.regressors import BernoulliMLPRegressor
 from tests.fixtures import TfGraphTestCase
 
 
@@ -75,8 +75,8 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_fit_normalized(self, input_shape, output_dim):
-        bmr = BernoulliMLPRegressorWithModel(
-            input_shape=input_shape, output_dim=output_dim)
+        bmr = BernoulliMLPRegressor(input_shape=input_shape,
+                                    output_dim=output_dim)
 
         observations, returns = get_train_data(input_shape, output_dim)
 
@@ -104,10 +104,9 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_fit_unnormalized(self, input_shape, output_dim):
-        bmr = BernoulliMLPRegressorWithModel(
-            input_shape=input_shape,
-            output_dim=output_dim,
-            normalize_inputs=False)
+        bmr = BernoulliMLPRegressor(input_shape=input_shape,
+                                    output_dim=output_dim,
+                                    normalize_inputs=False)
 
         observations, returns = get_train_data(input_shape, output_dim)
 
@@ -136,10 +135,9 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_fit_with_no_trust_region(self, input_shape, output_dim):
-        bmr = BernoulliMLPRegressorWithModel(
-            input_shape=input_shape,
-            output_dim=output_dim,
-            use_trust_region=False)
+        bmr = BernoulliMLPRegressor(input_shape=input_shape,
+                                    output_dim=output_dim,
+                                    use_trust_region=False)
 
         observations, returns = get_train_data(input_shape, output_dim)
 
@@ -168,12 +166,13 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
     ])
     # yapf: enable
     def test_log_likelihood_sym(self, output_dim, input_shape):
-        bmr = BernoulliMLPRegressorWithModel(
-            input_shape=(input_shape[1], ), output_dim=output_dim)
+        bmr = BernoulliMLPRegressor(input_shape=(input_shape[1], ),
+                                    output_dim=output_dim)
 
         new_xs_var = tf.compat.v1.placeholder(tf.float32, input_shape)
-        new_ys_var = tf.compat.v1.placeholder(
-            dtype=tf.float32, name='ys', shape=(None, output_dim))
+        new_ys_var = tf.compat.v1.placeholder(dtype=tf.float32,
+                                              name='ys',
+                                              shape=(None, output_dim))
 
         data = np.full(input_shape, 0.5)
         one_hot_label = np.zeros((input_shape[0], output_dim))
@@ -184,11 +183,11 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
 
         outputs = bmr.log_likelihood_sym(new_xs_var, new_ys_var, name='ll_sym')
 
-        ll_from_sym = self.sess.run(
-            outputs, feed_dict={
-                new_xs_var: data,
-                new_ys_var: one_hot_label
-            })
+        ll_from_sym = self.sess.run(outputs,
+                                    feed_dict={
+                                        new_xs_var: data,
+                                        new_ys_var: one_hot_label
+                                    })
 
         assert np.allclose(ll, ll_from_sym, rtol=0, atol=1e-5)
 
@@ -201,14 +200,13 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
     def test_optimizer_args(self, mock_cg, mock_lbfgs):
         lbfgs_args = dict(max_opt_itr=25)
         cg_args = dict(cg_iters=15)
-        bmr = BernoulliMLPRegressorWithModel(
-            input_shape=(1, ),
-            output_dim=2,
-            optimizer=LbfgsOptimizer,
-            optimizer_args=lbfgs_args,
-            tr_optimizer=ConjugateGradientOptimizer,
-            tr_optimizer_args=cg_args,
-            use_trust_region=True)
+        bmr = BernoulliMLPRegressor(input_shape=(1, ),
+                                    output_dim=2,
+                                    optimizer=LbfgsOptimizer,
+                                    optimizer_args=lbfgs_args,
+                                    tr_optimizer=ConjugateGradientOptimizer,
+                                    tr_optimizer_args=cg_args,
+                                    use_trust_region=True)
 
         assert mock_lbfgs.return_value is bmr._optimizer
         assert mock_cg.return_value is bmr._tr_optimizer
@@ -217,11 +215,10 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
         mock_cg.assert_called_with(cg_iters=15)
 
     def test_is_pickleable(self):
-        bmr = BernoulliMLPRegressorWithModel(input_shape=(1, ), output_dim=2)
+        bmr = BernoulliMLPRegressor(input_shape=(1, ), output_dim=2)
 
         with tf.compat.v1.variable_scope(
-                'BernoulliMLPRegressorWithModel/NormalizedInputMLPModel',
-                reuse=True):
+                'BernoulliMLPRegressor/NormalizedInputMLPModel', reuse=True):
             bias = tf.compat.v1.get_variable('mlp/hidden_0/bias')
         bias.load(tf.ones_like(bias).eval())
         bias1 = bias.eval()
@@ -235,18 +232,17 @@ class TestBernoulliMLPRegressorWithModel(TfGraphTestCase):
             assert np.array_equal(result1, result2)
 
             with tf.compat.v1.variable_scope(
-                    'BernoulliMLPRegressorWithModel/NormalizedInputMLPModel',
+                    'BernoulliMLPRegressor/NormalizedInputMLPModel',
                     reuse=True):
                 bias2 = tf.compat.v1.get_variable('mlp/hidden_0/bias').eval()
 
             assert np.array_equal(bias1, bias2)
 
     def test_is_pickleable2(self):
-        bmr = BernoulliMLPRegressorWithModel(input_shape=(1, ), output_dim=2)
+        bmr = BernoulliMLPRegressor(input_shape=(1, ), output_dim=2)
 
         with tf.compat.v1.variable_scope(
-                'BernoulliMLPRegressorWithModel/NormalizedInputMLPModel',
-                reuse=True):
+                'BernoulliMLPRegressor/NormalizedInputMLPModel', reuse=True):
             x_mean = tf.compat.v1.get_variable('normalized_vars/x_mean')
         x_mean.load(tf.ones_like(x_mean).eval())
         x1 = bmr.model.networks['default'].x_mean.eval()
